+++ src/language/quest.ts
@@ -22,6 +22,8 @@ enum QuestBlock {
  */
 export class Quest {
     
+    private static readonly quests = new Map<string, Quest>();
+
     /**
      * The block that holds quest directives.
      */
@@ -37,7 +39,9 @@ export class Quest {
      */
     public readonly qbn = new Qbn();
 
-    public constructor(public readonly document: vscode.TextDocument) {
+    private readonly version: number;
+
+    private constructor(public readonly document: vscode.TextDocument) {
 
         let block = QuestBlock.Preamble;
         this.preamble.start = 0;
@@ -79,6 +83,8 @@ export class Quest {
         this.preamble.end = this.qrc.start ? this.qrc.start - 1 : undefined;
         this.qrc.end = this.qbn.start ? this.qbn.start - 1 : undefined;
         this.qbn.end = this.document.lineCount - 1;
+
+        this.version = document.version;
     }
 
     /**
@@ -92,5 +98,28 @@ export class Quest {
             block = block.range || new vscode.Range(0, 0, 0, 0);
         }
         return new vscode.Location(this.document.uri, block);
+    }
+
+    /**
+     * Gets a `Quest` instance for the given document.
+     * @param document A text document with a quest to be parsed.
+     */
+    public static get(document: vscode.TextDocument): Quest {
+        let quest = Quest.quests.get(document.uri.fsPath);
+        if (!quest || document.version > quest.version) {
+            Quest.quests.set(document.uri.fsPath, quest = new Quest(document));
+        }
+
+        return quest;
+    }
+
+    /**
+     * Deletes the stored `Quest` instance associated to the given document.
+     * @param document A document whose corresponding quest is to be deleted.
+     */
+    public static delete(document: vscode.TextDocument): void {
+        if (Quest.quests.has(document.uri.fsPath)) {
+            Quest.quests.delete(document.uri.fsPath);
+        }
     }
 }